Add an executable that prints beats
This commit adds an executable to the package that prints the current
system time in beats and then exits.

The binary takes no arguments (there aren't many options with .beats
time, which is sort of it's beauty 🙂), but will print a usage message
and exit if any are passed. It exits with exit code 0 if run as `beats
-h` or with exit code 1 otherwise.